refactor(http-client-java): move TypeSpecFluentNamer into generator package
Relocate TypeSpecFluentNamer next to TypeSpecFluentPlugin and remove the leftover single-class fluent subpackage. Pure move, no behavior change.
